fixed remove post, updated notification functions

+++ app/src/main/java/com/rozdoum/socialcomponents/model/RecordingItem.java
@@ -39,7 +39,7 @@ public class RecordingItem implements Parcelable {
         return mLength;
     }
 
-    public void setLength(int length) {
+    public void setLength(long length) {
         mLength = length;
     }
